remove old navigation
